move btn add to menu
minor change of removing logout button from welcome card and adding  a such to a menu, for now. may look at it further just if there is. a case I may or not opt to modify this any further say to remove welcome page and do a dashboard or what not.., not sure as this was pretty much intended to be basically as is.